ci: add GitHub actions based CI
Also commit a package-lock.json file for reproducible local development.

As part of this we also amend the webpack config to drop the explicit
use of the webpack.optimize.UglifyJsPlugin plugin. This is implied by
the -p flag. Removing this from the webpack config, and also removing
the -p from the package.json script declaration means that we can
locally run:

    npm run build

in order to get a non-minified version for easy debugging.

Add a .gitattributes file to indicate dist is generated output for good
measure.